Moved utils to utils module
