+++ packages/service/config/jest.setup.ts
@@ -1,2 +1,2 @@
 import 'reflect-metadata';
-import '@service/jest-utils';
+import '@service/utils/jest-utils';
