+++ packages/service/src/domain/commands/processors/help-command.ts
@@ -4,7 +4,7 @@ import CommandProcessor from '@service/domain/commands/model/command-processor';
 import CommandRegistryService from '@service/domain/commands/registry/command-registry-service';
 import RegisterCommand from '@service/domain/commands/registry/register-command';
 import Configuration from '@service/infrastructure/configuration';
-import { notNull } from '@service/utils';
+import { notNull } from '@service/utils/utils';
 import { Service } from 'typedi';
 
 @RegisterCommand
